Remove repeated functions
