+++ tests/test_regression/test_rotation_and_derotation.py
@@ -4,7 +4,7 @@ import numpy as np
 from PIL import Image
 
 from derotation.derotate_by_line import derotate_an_image_array_line_by_line
-from tests.test_regression.recreate_target.rotation_and_derotation import (
+from tests.test_regression.recreate_target.shared import (
     load_rotated_images,
 )
